Fix the basic-rust client template for CLI's init command (#3646)
# Description of Changes

This template was never properly tested, and it contains wrong API usage
that can't even compile.

The only problem is that in order to make it work the person table has
to be public, and I'm not sure if we want to make that change.
Alternatively we could make basic-rust code not subscribe to anything
and just connect to the database.

# Expected complexity level and risk

2

# Testing

- [x] I compiled the template and checked that it works against

Co-authored-by: Zeke Foppa <196249+bfops@users.noreply.github.com>