commitlog: Change default max-records-in-commit to 1
It is almost always wrong / undesirable to pack more than one
transaction in a commit, so adjust the default accordingly.

This also avoids surprises when using `#[serde(default)]` with nested
structs -- serde evaluates the default depth-first, so overriding a
single field in a nested struct will not consider any
`#[serde(default = "custom_default")]` annotations on the parent.