Add procedure HTTP request API
With this PR, procedures (at least, those defined in Rust modules) can perform HTTP requests!
This is performed through a new field on the `ProcedureContext`, `http: HttpClient`,
which has a method `send` for sending an `http::Request`, as well as a convenience wrapper `get`.

Internally, these methods hit the `procedure_http_request` ABI call / host function,
which uses reqwest to perform an HTTP request.
The request is run with a user-configurable timeout which defaults and is clamped to 500 ms.
Rather than exposing the HTTP stream to modules,
we download the entire response body immediately, within the same timeout.

I've added an example usage of `get` to `module-test` which performs a request against `localhost:3000`
to read its own schema/moduledef.

Not included in this commit, but intended to be within this PR, are:
- [ ] A test using the `sdk-test` framework.
- [ ] Expanded documentation.

Left as TODOs are:
- Metrics for recording request and response size.
- Improving performance by stashing a long-lived `reqwest::Client` someplace.
  Currently we build a new `Client` for each request.
- Improving performance (possibly) by passing the request-future to the global tokio executor
  rather than running it on the single-threaded database executor.