Add `spacetime_worker_reducer_returned_errors_total` metric (#3613)
# Description of Changes

Currently, we have a metric for reducer panics called
`spacetime_worker_wasm_instance_errors_total`. This commit adds a metric
for tracking errors returned from the module, like for example an Err
result in Rust, or throwing a SenderError in TypeScript

This will be needed for the web database overview dashboard, but it
should be useful for tracking sender errors in general.

# Expected complexity level and risk

1

# Testing

- [x] I've tested the change locally verifying the counter increments
when an instance returns an error

---------

Co-authored-by: Tyler Cloutier <cloutiertyler@users.noreply.github.com>