Delete duplicated docs folders
Somehow (apparently after #3494) we wound up with a bunch of duplicate folders in our repo,
both the old title-case names and the new lower-case names.
All of the remaining files in the title-case directories
have their last change in #3343, well before #3494,
so I don't believe we're losing any intermediate changes by deleting them.
I'm not sure how this happened, but it seems to be an easy fix.