Update typescript package size limits (#3786)
# Description of Changes

<!-- Please describe your change, mention any related tickets, and so on
here. -->

I got these errors when I tried sending out the typescript NPM package:

```
  cjs (brotli)
  Package size limit has exceeded by 165 B
  Size limit: 26 kB
  Size:       26.16 kB brotlied
  
  esm (brotli)
  Size limit: 26 kB
  Size:       25.87 kB brotlied
  
  esm (gzip)
  Package size limit has exceeded by 536 B
  Size limit: 30 kB
  Size:       30.54 kB gzipped
  
  esm (uncompressed)
  Package size limit has exceeded by 3.77 kB
  Size limit: 160 kB
  Size:       163.77 kB
  
  esm min (brotli)
  Package size limit has exceeded by 194 B
  Size limit: 14 kB
  Size:       14.19 kB brotlied
  
  esm min (gzip)
  Package size limit has exceeded by 144 B
  Size limit: 16 kB
  Size:       16.14 kB gzipped
  
  esm min (uncompressed)
  Package size limit has exceeded by 1.29 kB
  Size limit: 65 kB
  Size:       66.29 kB
```

# API and ABI breaking changes

None

<!-- If this is an API or ABI breaking change, please apply the
corresponding GitHub label. -->

# Expected complexity level and risk

1

<!--
How complicated do you think these changes are? Grade on a scale from 1
to 5,
where 1 is a trivial change, and 5 is a deep-reaching and complex
change.

This complexity rating applies not only to the complexity apparent in
the diff,
but also to its interactions with existing and future code.

If you answered more than a 2, explain what is complex about the PR,
and what other components it interacts with in potentially concerning
ways. -->

# Testing

<!-- Describe any testing you've done, and any testing you'd like your
reviewers to do,
so that you're confident that all the changes work as expected! -->

- [x] After this change I was able to publish the typescript NPM
package.