Actually report reducer fuel used (#3799)
# Description of Changes

In the `call_reducer_with_tx` function we only reported the WASM fuel
used by view evaulation, but not the stats from the actual reducer call.
This PR fixes it and we now properly record it.

# Expected complexity level and risk

1

# Testing

I've tested the change locally. Before the change the reported metrics
were always zero after running any reducer. Now the usage is reported
properly.