Implement DbContext for AnonymousViewContext and ViewContext (#3787)
# Description of Changes

The trait has already been implemented for ReducerContext and TxContext,
but implementation for AnonymousViewContext and ViewContext is still
missing.

## Background why I need this change

In SpacetimeDSL, I currently have the ReducerContext hard-coded as a
prerequisite for creating a DSL instance (see
https://github.com/tamaro-skaljic/SpacetimeDSL/blob/main/src/lib.rs#L9-L33)
.

With the addition of

- ViewContext,
- AnonymousViewContext,
- ProcedureContext, and
- TxContext,

I am now faced with the dilemma that I also have to support SpacetimeDSL
with the other contexts.

I hope to be able to use `spacetimedb::DbContext<DbView =
spacetimedb::Local>` and `spacetimedb::DbContext<DbView =
spacetimedb::LocalReadOnly>` as a common denominator.

(I signed the CLA)

# API and ABI breaking changes

No

# Expected complexity level and risk

1

Signed-off-by: Tamaro Skaljic <49238587+tamaro-skaljic@users.noreply.github.com>
Co-authored-by: joshua-spacetime <josh@clockworklabs.io>