Don't error during replay when deleting a table with no in-memory instance
Prior to this commit, we treated it as an error
(in fact a panic, prior to an earlier commit of mine today)
when deleting a table during replay which didn't have an in-memory instace.
This should not have been an error, as it's not problematic,
and it can occur in valid commitlogs when a table is initially empty
and is never used prior to its deletion.

With this commit, don't error.